requirements.txt
pyelftools==0.32
pytest==8.4.0
